Refactored with sieve
Need to fix invalid pointer bug

+++ include/problem3.h
@@ -7,9 +7,9 @@ extern unsigned long int INPUT;
 extern int TEST_INPUT;
 extern int TEST_OUTPUT;
 
-int problem3_solution(unsigned long num);
-int largest_prime_factor_of(int num);
-int *sieve_of_eratosthenes(int highest_num);
-bool is_prime_number(unsigned long num_to_test);
+int problem3_solution(unsigned long long num);
+long long int largest_prime_factor_of(long long int num);
+long long int *sieve_of_eratosthenes(long long int highest_num);
+bool is_prime_number(unsigned long long int num_to_test);
 
 #endif
